Bug#29913928: ER_SLAVE_HEARTBEAT_FAILURE PRINTS BOGUS DATA IN SERVER ERROR LOG
Problem:

When the server print a ER_replica_HEARTBEAT_FAILURE error message, it is not
respecting the length of the source log file name, leading to print bogus
data (it might also become a leak) in Heartbeat Version 1 event.

Also, when the replica detects a heartbeat event is not valid, it is trying to
print the "invalid" data in the error log. As invalid events may be
corrupted, the server shall never use its contents.

Fix:

Deserialize the source log file name in a way so that we allocate memory for the
file name, copy it from the event raw buffer and adds the \0 at the end of the content.

Also, when replica receives an invalid heartbeat event, it should not print its
content to the error log.

Change-Id: I3b7c44a4939f31bc341741a4d3bfc4b2882e4576