Bug#36135466 test_mgmd fails with suite/ndb/tls.cnf
The fix for bug#35696649 added support for a large set of NDB
tests and tools to run with TLS enabled. That patch also added
the "extern" references to command-line option settings
in NdbMgmd.hpp. It is the implicit (and surprising) use of
these variables that causes the NdbdWithCertificate test to
fail in a TLS environment.

Remove the extern references, and require the caller to
explicitly set TLS options prior to NdbMgmd::connect().

Change-Id: Id52f9af6cc404e0b2ae68a5bdd9baef7eaae8930