BUG#36191370 refused discover in transaction causes rollback
Problem:
Implicit rollback when refusing to discover table in an active
transaction causes entire transaction to rollback.

Analysis:
When table definition mismatch occurs between NDB and MySQL dictionary
the table will be installed into MySQL DD from NDB, this is called
discover. Installing into DD requires committing the current active
transaction and thus discover is not possible while in an active
transaction, discover need to be refused and an error and warning
returned to the user. Unfortunately the code checking for being in an
active transaction also checked if the table already existed in DD thus
implicitly rolled back the transaction. This meant that a subsequent
read from same table in the same transaction would
allow discover.

Solution:
1. Remove check whether or not table already exists in DD, instead
always
refuse discover while in transaction. This means that also subsequent
attempts to use the same table, in the transaction, will return the same
error.
2. Improve warning message to describe to user that discover was refused
due to being in active transaction.
3. Use correct external return code ER_TABLE_DEF_CHANGED rather than
internal HA_ERR_TABLE_DEF_CHANGED when formatting the error message.
4. Update existing tests to:
 a) accept that it's not possible to discover
any table while in an active transaction
 b accept the new error code

Change-Id: Ie1bd877253c9895229691d2c154f73329ee6220d