Bug#36108397 Upgrade to latest protobuf library
Post-push fix:
routertest_fuzz_http_auth_backend links in libmysqlxclient_lite.a
and if this links PUBLIC with libmysqlxmessages_lite.a we get ODR violations.

routertest_fuzz_http_auth_backend links with its own
   libmysqlrouter_mysqlxmessages.so.1
in order to get the necessary symbols from libmysqlxmessages_lite.a

The fix is to link mysqlxclient_lite with ext::libprotobuf-lite instead.
This pulls in the necessary symbols when building on Windows,
and avoids the ODR violations for the router executable.

Change-Id: I8301414c8895735b97ea66818adbe78cb5f749e5