Bug#35855294 Inconsistent results of SELECT statement with EXISTS clause - part III: row subquery with outer reference
Scalar row subqueries which are part of an expression with window
functions can never be executed early, as the result can not be
encoded in a single field to go onto the input table for the windowing
step.  As such, we must ensure they can be executed after the result
of the window function is ready. In this patch, we ensure that any
outer references in such scalar row subqueries point to the window
output table where their correct values are to be found.

The test file has a new query that used to get the wrong result before
this patch.

Change-Id: Id53d42c26c6d716189034e72ed0138d112fa4645