Bug#36190724 Make test perfschema.dd_version_check more constraining to detect invalid PFS_DD_VERSION changes
Context:

The existing test case perfschema.dd_version_check is not
a regular functional test.

It executes logic to detect broken upgrades,
to enforce that PFS_DD_VERSION is properly maintained,
when implementing performance_schema features.

Problem:

Before this change, it was possible to:
- make table changes in the performance_schema
- adjust test perfschema.dd_version_check with a new signature,
  to make the test pass,

but forget to adjust the global variable PFS_DD_VERSION.

The consequences of forgetting to adjust PFS_DD_VERSION
are severe, and causes bugs in the upgrade process,
because different releases of MySQL, using different DDL,
claim to have the same version number in the data dictionary,
stored on disk.

Solution:

This test, perfschema.dd_version_check, is now enforcing more consistency
checks, to detect if PFS_DD_VERSION was effectively updated.

Also, awareness to the maintainer comments located in the test
has been improved.

Change-Id: Iab3290a0110a170cec7abedd9d0073cf52c15b03