WL#16098: Introduction of Named Time Zone in HEATWAVE.
This work log introduces support for the Named time zone during
the load and execution of queries in HEATWAVE.
The time-zone offset is determined during the cluster setup by
reading the operating system's zone info files.

In addition to the time-zone session variable, support for the
time zone has been extended to the following primitives:
CONVERT_TZ(), CURRENT_TIMESTAMP(), CURRENT_TIME(),
CURRENT_DATE(), UNIX_TIMESTAMP(), and FROM_UNIXTIME().

The behavior of the time zone in HEATWAVE is anticipated to be
identical to MySQL, with the only distinction being the method
of obtaining the offset. In MySQL, the zoneinfo files are read
only once, and the offset information is stored in the
mysql.time_zone* tables. In Heatwave, the offset is read from
the operating system on every restart.

The remaining functionality and limitations of the time zone in
HEATWAVE are expected to be consistent with MySQL.

Change-Id: I87c1cd62b4131021f9aa1ac7d183d9be47ba4279