Bug#36099491: Assertion secondary_engine_cost_hook != nullptr failed with hypergraph_optimizer [1/2, lateral]
In some queries with lateral derived tables or derived tables with
outer references, no valid join order was found by the hypergraph
optimizer when running with the optimizer_max_subgraph_pairs system
variable set to a very low value.

The problem was that the graph simplification, which was triggered
because of the low optimizer_max_subgraph_pairs value, didn't take
lateral dependencies into account, and didn't hesitate to rewrite the
hypergraph so that the derived table had to be joined first with a
table on the outer side of a LEFT JOIN which had the lateral
dependencies of the derived table on the inner side. Now the
hypergraph said the derived table had to be to the left of the lateral
dependencies, whereas the lateral dependencies should force it on the
right side. This made it impossible to come up with a valid plan.

Fixed by making GraphIsJoinable() take lateral dependencies into
account, so that the graph simplification detects that such
restrictions make it impossible to find a valid join order and doesn't
try to rewrite the hypergraph this way.

Also, IWYU was run to clean up #include statements in the files
touched by the patch.

Change-Id: Ic7541d314165e621a131a1f27bad802c3e59554e