Bug#36356532 testBackup -n  Bug17882305 leaves Error Inserted after running
Test injects error 10046 in Backup instances to force DIH to change the
next fragment to scan.
Error is injected to Backup Proxy and then proxy resends the error
to all workers where it is eventually cleared. But, in Backup proxy
instance error is never cleared.
Fix:
Clear error in test side (Insert Error 0) to force it to be cleared
in all Backup instances (including Backup proxy).

Change-Id: Ieafdde52bf19c6d3946d6d53dd26eab9dc329526