WL#14826 Concurrent Query Execution [ patch 5 ] Bug#36060305 - wl14826 : query takes 1hr 33 mins to return OOM with 1T on 2 and 3 nodes Bug#36055769 - wl14826:query taking more than 45mins to return OOM
This worklog enables RAPID to run queries concurrently instead of one
after the other. Specifically, it enables the plugin to offload multiple
queries, RAPID servers to concurrently process multiple queries and
mechanisms for reducing the chance of OoM due to memory pressure from
concurrent queries. This patch introduces mechanisms for reducing the
probability of an OoM and for handling OoMs when they happen using
retries.

Change
======
-Memory-based admission: use memory estimates to decide which queries
should be executed concurrently. The sum of their estimates needs to
be less than the total memory.

-Retry mechanism: implement retry mechanism for query failures due to
out-of-memory. The retry mechanism exploits memory-based admission.
Every time a query is retried, it requires more tickets so concurrency
is restricted. Eventually, memory-intensive queries run on their own
and if they fail then, they give out-of-memory error.

-Memory-aware scheduler: the scheduler uses the estimates to predict
whether the concurrent queries will get an OoM error. It then takes
action to risk the execution only for queries that have already
exceeded their estimates. The decision is communicated to other
nodes and cores to avoid deadlocks due to different decisions.

Bug 36060305:
Bug 36055769:

Some TPC-H queries took a long time before they return OOM. This is
due to runtime scheduler threads staying active while queries are
freeing memory. As the specific queries took a significant amount
of time in single node-single core sorting, all but one cores were
left busy-waiting, affecting memory trim performance.

This patch makes runtime scheduler threads sleep except for the
case that there are queries that are ready to be scheduled but
are waiting for the memory of finishing queries to be released.

Change-Id: I63d280cf19d96e0128c62dba14a19236ccdc0e37