BUG#36199962: show_all_binlogs.inc overwrites $binlog_position
* Background

Many tests read a value into $binlog_position in one position and use
it at a later position.

show_all_binlogs.inc is used for debugging. You typically source it in
some places in a test to observe the binlog contents at those times.

* Problem

show_all_binlogs.inc modifies $binlog_position. So if it is used in a
section where the test expects to have saved a particular value in
$binlog_position, it interferes with the test. This makes debugging
harder.

* Fix

Use $_sab_binlog_position instead of $binlog_position in
show_all_binlogs.inc. "sab" is the abbreviation of "show_all_binlogs".
This convention is used in many .inc files to minimize global
namespace pollution.

Similar for 3 other variables in this file.

Change-Id: I85abefbddc1f5bca1ab654604442c0effab2a301