Bug#36191538: Make Item_func_make_set adhere to Item_func argument style
Class Item_func_make_set stores most of its arguments in the "args"
array supplied by the base class Item_func. However, the first argument
is stored in a dedicated member "item". This anomaly means that a number
of standard functions implemented by Item_func cannot be used by this
class.

This is fixed by removing the "item" member and instead carrying it as
args[0] and shifting other arguments by one place. As a consequence,
several special functions can be removed.

Note that a special fix_fields() implementation is still required to
preserve existing nullability.

One Rapid test is changed due to this. This is because of missing
specific implementation of compile() function in existing code.

Change-Id: I62476d019a87caaad14de843fb7d191aff80aa3e