Bug#36235788 - GET_LOCK causes Query Termination error
GET_LOCK() waits for a user-level lock to be freed if it is
already acquired by another user session. In a wait loop for
GET_LOCK(), the connection state is observed to ensure that the
client session still exists. If the connection is gone, the wait
loop is exited with an error. To check the connection state, the
current THD protocol's 'connection_alive()' is invoked. However,
Protocol_local_v2's connection_alive() always returns false.
Therefore, when the cursor is enabled and GET_LOCK() is invoked
from the MLE SP, the wait loop is terminated, and an error is reported.

GET_LOCK() call from MLE SP works as expected without cursors.

To resolve this issue, the Protocol_local_v2's connection_alive()
method is modified to return true if the user connection is bound
to the THD (session).

Change-Id: Ic487173a6e1e4917bfabc9e2641fd3cc73fcd4c7