Bug#36172350: Wrong result with QUALIFY referencing columns               not present in the SELECT list.
If a column referenced in QUALIFY condition is not present in the SELECT
expressions list and if the condition does not have any window functions
or aggregate functions, then we get wrong results. This is because the
column is not added to the field's list for it to have the correct value
when the condition gets evaluated. Usually such expressions get added to
the list when a call is made to split_sum_func(). In this case since
there are no window functions or aggregate functions in the expression,
split_sum_func() never gets called.
Solution is to add the field to the SELECT expression list when a column
in QUALIFY condition is resolved in the FROM clause instead of the SELECT
list.

Change-Id: Ie9b61b21d39fdedfb6f34381da06e6f548367db6