Bug#36200398: Wrong empty result set with hypergraph, windowing and derived tables
The hypergraph optimizer returned wrong results for some queries that
used window functions that had references to expressions from merged
derived tables in their PARTITION BY or ORDER BY clause.

The problem was that the hypergraph optimizer made a mistake when
rewriting the item tree in PARTITION BY or ORDER BY to redirect it to
the temporary table fields that holds the materialized expression.
Instead of replacing the entire expression with a temporary table
field, it kept the expression and instead rewrote the arguments of the
expression to point to temporary table fields. Since the item
representing the expression was used other places too, rewriting its
arguments disturbed the other uses of it.

In particular, the condition in a filter representing the WHERE clause
was rewritten to point to fields in a temporary table that was not
available before windowing. Because of this, the filter always
evaluated to false, and no rows were returned.

The cause of the bug was that the hypergraph optimizer used
Item::real_item() to peel off the Item_ref that the resolver puts on
top of every item in Window::m_partition_items and
Window::m_order_by_items, but real_item() works recursively and can
remove multiple layers of Item_ref wrappers. In the problematic
queries, the Item_ref contained an Item_view_ref, thanks to the merged
derived table. Since the hypergraph optimizer didn't see the
Item_view_ref, but only its child, it did not see that it had made a
temporary table column for that Item_view_ref and could replace it.

Fixed by calling Item_ref::ref_item() instead of Item::real_item(), in
order to peel off exactly one Item_ref.

Also clean up #include statements in the modified files, following the
suggestions from the IWYU tool.

Change-Id: I94323cbee62b1c12da004906dff0c40a4b1e926c