Bug# 35801548 - Ndb System Restart failures risk recoverability
In a System Restart (SR), there will not be any transaction ongoing
until the restart is completed. Since all nodes are starting from
their local checkpoints (LCP), there is no need for CopyFrag phase
where a starting node synchronises with a running node and thus the
possibility of undo log capacity being exceeded during restart will
not occur.

After partial LCP is introduced in 7.6, one local LCP is performed
during early start phase in order to make partial LCPs run faster when
distributed LCP starts.

Theoretically, for an SR, it should not be necessary for on-disk state
to be damaged as there is no CopyFrag state or local partial LCPs are
required.

When a GCP completes during a node restart, DIH sends GCP_SAVEREQ to
LQH. LQH then requests CNTR to write the new gci as max restorable gci
to local SYSFILE together with the flag
'NODE_NOT_RESTORABLE_ON_ITS_OWN' in SYSFILE. This continues until
restart completes and the new GCI is written together with
'NODE_RESTORABLE_ON_ITS_OWN' flag. No newer gci will be written to
local sysfil after this point.

Problem: During an SR, any node failure when SYSFILE is in the state
'NODE_NOT_RESTORE_ON_ITS_OWN' makes it impossible to get the system up
again. This occurs between the node status STARTING (from Last
completed phase 4 to Last completed phase 8) and STARTED, as printed
by the mgm client when the cluster node status is queried.

Solution: Since the on-disk state is not damaged during an SR, it is
not necessary to set the flag 'NOT_RESTORABLE_ON_ITS_OWN' before the
first LCP starts, as it is required for a node restart performing CopyFrag.

Reviewed by: frazer clement <frazer.clement@oracle.com>

Change-Id: Ifc439fb484384bfae8e9900075e1fe255a757449