Bug#35671897 Concurrent Binlog rotate and DDL commit hit safe_mutex assertion
Problem:
Handling an incident while concurrent transactions are committed to the
binlog cause eternal MySQL Server lockup.

Analysis:
An incident occurs in one of two ways:
1. When a transaction which can't be rolled back fails to write to the
binlog cache.
2. When a session want to record a fatal error from another subsystem.

The incident will write an Incident_log_event which contains message and
error code that describes the problem. The Incident_log_event should
always be preceded by a Gtid_log_event, this makes it possible for
binlog consumer to skip over the incident in same way as other
statements/transactions. After having written the Incident_log_event,
the binlog should be rotated and checked for purge.
The reason for always rotating and checking the log for purge is that it
will potentially free up disk space and thus solve the cause of the
incident.

The main reason for the lockup is that the session rotates and purges
the binary log from outside the binlog group commit (BGC), while other
concurrent sessions may potentially be in the middle of an ongoing
binlog flush.

The normal procedure is that the binary log is written in the flush
stage of BGC, and rotations/purges at the end of the commit stage of
BGC. This is known to work and well tested; the concurrent use of
another execution flow was the reason for the bug. Therefore we fix the
bug by following this normal procedure also for incidents.

Solution:

Defer handling of incident until inside binlog group commit (BGC). The
flush stage will check whether any binlog cache is marked as having an
incident, in such case it will replace the content of the binlog
cache(s) with an Incident_log_event which will then be written to the
binlog. As part of handling the incident it will also request that
binlog rotation and old file purge logic are performed when that stage
is reached. In addition the Incident_log_event will cause dependency
tracker to mark all subsequent transactions to depend on it.

The patch has following changes:
1. The existing interface to set an incident on the binlog cache is
extended to take a message describing the incident, this is used by
session threads that experience failure writing to binlog cache.

2. The interface for recording an fatal error in another subsystem is
changed with a new function that marks the binlog cache as having an
incident and then commits it.

3. The patch slightly change behaviour so that an incident will always
rotate the binlog and execute the purge logic, this was not the case
earlier when failing to write to the binlog cache during commit - in all
other cases it would rotate and purge. This is done by extending the
Binlog_group_commit_ctx with an indicator of whether binlog max size was
seen to be exceeded during flush and an indicator whether a session want
to force rotate. These members are then aggregated over all committed
sessions before the binlog rotate+purge commence at end of BGC. This new
mechanism is using a somewhat branchless strategy.

Testing:
New tests are added which provoked the lockup before fix.
Existing tests are modified to test that binlog are always rotated and
adapt to slight changes in error message which is now more generically
phrased.
The test for checking functionality handling the case when binlog cache
does not exist are removed since it was proven that such case did not
exist - neither in old or new implementation.
The upgrade compatibility test which intentionally writes an
Incident_log_event without preceding Gtid is modified to use a
more generic DBUG point.

Change-Id: I731727242c40227ffe816fc4f86af59b5520936f