Bug#36221823 SQL grammar has four needless shift/reduce conflicts
The rule
    opt_field_length:
           %empty { $$= nullptr; /* use default length */ }
         | field_length
         ;
where
   field_length:
          '(' LONG_NUM ')'      { $$= $2.str; }
        | '(' ULONGLONG_NUM ')' { $$= $2.str; }
        | '(' DECIMAL_NUM ')'   { $$= $2.str; }
        | '(' NUM ')'           { $$= $2.str; };

gives rise to shift/reduce conflicts.

The patch adds a %prec directive to the empty alternative of the
opt_field_length production, which brings the number of shift/reduce
conflicts down to 59. This is done by using (and renaming since it is
now being used in more than one type of context) the artificial token
SUBQUERY_AS_EXPR (renamed to PREFER_PARENS).

Change-Id: I7808d579db18598c03e59398b222b19ce20b3bef