Bug#36357373 testDict leaves Error 5076 Inserted after running
Error is injected to LQH Proxy and then proxy resends the error
to all workers where it is eventually cleared. But, in LQH proxy
instance error is never cleared.
Fix:
Clear error in test side (Insert Error 0) to force it to be cleared
in all LQH instances (including LQH proxy).

Change-Id: Icb06ee1b360fa424a1c972ec799749891b2f0f22