Bug#36235709 - Mysqld crashes if an SQL-callout query is executed in finally clause in MLE SP.
The Statement Handle Interface resets the Diagnostics Area (DA)
before executing a statement. If an error occurs during the execution
of a statement, then successfully executing another statement using
the Statement Handle Interface will clear the DA. Returning failure
from the main statement with a clear DA results in assert conditions
to check the DA state in error case, leading to failure. The execution
of an MLE SP, which is calling a SQL callout in the finally block
of try..catch..finally, failed due to this reason.

To address this issue, a temporary Diagnostics Area (DA) instance
is now used to execute a statement using the Statement Handle
Interface when an error has already occurred. In non-error cases,
the current DA is reset before executing a statement.

With this fix, the behavior will be as follows:

1. If an error occurs, and the next statement is not executed, then
   error and warning messages from the main statement are stored
   in the main Diagnostics Area (DA).

2. If an error occurs, and the next statement (in the finally block)
   is executed successfully, then error and warning messages from
   the main statement are stored in the DA.

3. If an error occurs, and the next statement (in the finally block)
   also fails, then error and warning messages from the second
   statement are stored in the DA.

4. If an error occurs, and the next statement is executed successfully
   from the error handler (catch block), then the error from the
   main statement is cleared from the DA.

5. If an error occurs, and the next statement execution from the
   error handler (catch block) also fails, then error and warning
   messages from the second statement are stored in the DA.

Change-Id: I430098a284d22a5cac2d3940b68bc8abe9686d1a