Bug#36220640 Mrr implement runs much slower than before
Queries using the Multi-Range Read optimization run slower after the
fix for Bug#36017204.

The reason is that the insertion sort used to sort the records in the
MRR buffer does not perform well when the number of records in the
buffer is high.

Fixed by replacing the insertion sort with a call to qsort() in the C
standard library. It performs better on large buffers.

Change-Id: I39da5db5c329d482bf83b18fb27e52a0c5fcae03