Bug#36240027 test fails if mysqlrouter.conf exists
The AppTest.CmdLineExtraConfigNoDefaultFail test of
routertest_router_mysqlrouter_app fails when mysqlrouter.conf exists
in the current directory.

This does not happen in test-environments where the current directory
is under control and empty, but in build-environments where a the
build-dir may contain other fails.

The existing check in the test for "mysqlrouter.conf in current dir"
is faulty as it replaces {origin} with the program-name.

Change
======

- replace {origin} with the origin-dir, instead of program-name

Change-Id: I522cb199a90219eff4ab3f929a4c9a8232df7885