WL#16089: Remove the usage of the deprecated CHANGE MASTER command in tests and other components [Step 1]
This worklog removes the usage of CHANGE MASTER and its clauses in
tests files and other minor usages. In the process minor code and test
usages of the renamed terminology may cause some code and other test
variables or structures to be modernized as well.

Step 1: Remove occurrences of CHANGE MASTER and most of its clauses in MTR tests

Change-Id: I239f8abb38f10486a8ee82ceb509c7c08938ee8d