WL#15649 Expose Router configuration in metadata
The Router now exposes its static configuration in the metadata on
boostrap and updates it each time it starts. It is exposed as
a JSON in the routers.attributes.Configuration field.

Additionally on fist bootstrap of the given version, the Router exposes the
default bootstrap options for this version in the cluster/clusterset table
respectively, in the router_options.Configuration.<version>.Defaults field.

The options that the given version of the Router supports changing dynamically
via the metadata are exposed in
router_options.Configuration.<version>.ConfigurationChangesSchema  field.

Change-Id: I167622519c48eb1f60ef89400150e7ec74890ee1