Bug#36220663 Router using deprecated default_authentication_plugin variable
Since the mysql_native_password is deprecated and been removed, recently
the Router introduced a mechanism that tries to automatically upgrade the
Router user account that uses it. Part of this procedure was trying to figure
out the default_authentication_plugin using
"select @@default_authentication_plugin" query. This status variable is also
being removed now though.

This patch simplifies the automatic account upgrade procedure by removing the
step where the default plugin is queried. Instead the account is now always
being upgraded to use caching_sha2_password. It is up to the user to maunually
do the upgrade when this fails for some reason (like for example authenitcation
policy that forbids caching_sha2_password, etc.)

Change-Id: I1547f95fd41250bde37eda8b2eecdd2d448c4dd5