Bug#36225456 MySQL Router - Failed checking the Router account authentication plugin
When the Router is bootstrapped it checks if the metadata user is not using
unsupported mysql_native_password plugin. First part of this check is query
of the users table for the host and plugin. If this fails the procedure is
skipped but the user sees an error messsage as an bootstrap output. Meanwhile
this can be no error scenario. For example the user that is used to bootstrap
can have no privileges to query users table. This will be the case if the user
created by the Shell is used.

This patch removes the confusing error message from the bootstrap output.

Change-Id: Ic9509a57f5886747f2fe401e566b9f0e2a3c6bd5