Bug#36070647 unnesting scalar subquery containing count and having cond produces wrong result [2/2 bug fix]
A query like the following:

SELECT ( SELECT COUNT(1) AS cnt
         FROM t2
         WHERE t2.a = t1.a
         HAVING cnt > 0) FROM t1;

was transformed to

SELECT COALESCE(derived_1_2.cnt,0) AS cnt
FROM t1
     LEFT JOIN
     ( SELECT COUNT(1) AS cnt,
              t2.a AS a from t2
              GROUP BY t2.a
              HAVING (cnt > 0)) derived_1_2
     ON derived_1_2.a = t1.a

as per WL#13520 "A special case: COUNT". However, the presence of a
false HAVING condition in the subquery should semantically change the
correct result of the scalar subquery from zero to NULL. This is what
happens in the untransformed query, but not in the transformed case.

In this patch, we augment the transform to compensate for this, with
the following transform:

SELECT COALESCE(derived_1_2.cnt,
                IF(derived_1_2.having_expr IS NOT NULL, 0, NULL)) AS cnt
FROM t1
     LEFT JOIN
     ( SELECT COUNT(1) AS cnt,
              t2.a AS a,
              COUNT(1) > 0 AS having_expr
              FROM t2
              GROUP BY t2.a
              HAVING (cnt > 0)) derived_1_2
     ON derived_1_2.a = t1.a

This includes the HAVING condition in the select list of the derived
table, which allows the COALESCE to test for this and give the correct
result.

Change-Id: If362a1bf94c3ba253fb3babdd92e35fec55632ee