Bug#36245132 Router uses default auth plugin not checking it is password based.
When Router is bootstrapped with no --account parameter
it creates a new metadata user on the server.
The following query is used for that:
CREATE USER IF NOT EXISTS username@hostname IDENTIFIED BY 'password'

This can be problematic if the configured default authentication plugin on the
server is some NON-1-factor and/or NON-password-based authentication plugin.

This patch changes the query that creates an user to always force
caching_sha2_password plugin:

CREATE USER IF NOT EXISTS username@hostname IDENTIFIED WITH
                      `caching_sha2_password` BY 'password'

Change-Id: I95ef7b01c622a36bb9fdfed552a1b5b2caaa3c3c