Bug#34754817 Reconsider message "Failed to set ODirect errno:" as it is repeated
When ODirect is set in configuration the logging has changed.

First during startup data nodes probes whether ODirect is settable for
the four different root paths that can be configured FileSystemPath,
BackupDataDir, FileSystemPathDataFiles, FileSystemPathUndoFiles. The
probe is done by creating a file ndb_<nodeid>_prb.tmp and try to turn on
ODirect. The result will be logged whether it was succeded, failed to
probe (for example the root was a read only directory), or failed to set
ODirect.

Since the base paths can have other filesystems mounted in subpaths
setting ODirect will continue to be tried on for relevant files
independent of what the initial check reported.

Later when a file is opened the result of setting ODirect will be
recorded but not logged. The logging will happen around 4 hours apart
for each root path if there have been any failures. If there have been
no failure no logs will be emitted.

For table space files and logfile group files that are created with
absolute path failures will always be logged at once. Creation of such
files are rare, and since absolute paths are used they are likely not
related to a base path.

Also did some refactoring adding FsOpenReq::Versions enum.

Example of logging failure of initial probing for ODirect support on
base paths:

2024-03-01 18:47:19 [ndbd] WARNING  -- Failed to set ODirect for file /.../mysql_cluster.1/ndbd.1/ndb_1_fs/ndb_1_prb.tmp under FileSystemPath (errno: 22).
2024-03-01 18:47:20 [ndbd] WARNING  -- Failed to set ODirect for file /.../mysql_cluster.1/ndb_1_prb.tmp under BackupDataDir (errno: 22).
2024-03-01 18:47:20 [ndbd] WARNING  -- Failed to set ODirect for file /.../mysql_cluster.1/ndbd.1/ndb_1_fs/TS/ndb_1_prb.tmp under FileSystemPathDataFiles (errno: 22).
2024-03-01 18:47:20 [ndbd] WARNING  -- Failed to set ODirect for file /.../mysql_cluster.1/ndbd.1/ndb_1_fs/LG/ndb_1_prb.tmp under FileSystemPathUndoFiles (errno: 22).
2024-03-01 18:47:20 [ndbd] WARNING  -- ODirect configured but is not supported on any of the base paths. Was that intentional?

Example of 4 hour report of failing ODirect usage:

2024-03-01 22:47:20 [ndbd] WARNING  -- Setting ODirect have failed for 136 files and succeeded for 0 files under /.../mysql_cluster.1/ndbd.1/ndb_1_fs/ (FileSystemPath) since last warning.
2024-03-01 22:47:20 [ndbd] WARNING  -- Failed to set ODirect for file /.../mysql_cluster.1/ndbd.1/ndb_1_fs/LCP/0/T13F0.Data under FileSystemPath (errno: 0, block size 4096, alignment 512, direct io 1, avoid on append 0, io block size 0, alignment 0).
2024-03-01 22:47:20 [ndbd] WARNING  -- Setting ODirect have failed for 28 files and succeeded for 0 files under /.../mysql_cluster.1/ (BackupDataDir) since last warning.
2024-03-01 22:47:20 [ndbd] WARNING  -- Failed to set ODirect for file /.../mysql_cluster.1/BACKUP/BACKUP-14/BACKUP-14-PART-1-OF-4/BACKUP-14-0.1.Data under BackupDataDir (errno: 0, block size 4096, alignment 512, direct io 1, avoid on append 0, io block size 0, alignment 0).

Change-Id: I03de38d5e4f17a172e6db21daf0bc3d517b855bd