Bug#36246859: Collation issue: ERROR 1253 (42000): COLLATION ''               is not valid for CHARACTER SET
Condition pushdown to a view fails with a collation mismatch
if the view was created with a different charset than the
charset used when querying the view.
Problem is seen if the underlying fields in the view are
literals with COLLATE statements. The string literal with
the COLLATE statement is cloned when replacing expressions
in the condition that is being pushed down with the
expressions from the view. The string literal is currently
parsed with the connection charset which in this case is
different from the one that was used when the view was
created. Therefore the COLLATE statement fails.
Creation context for a view has the connection charset and
collation information which was used when the view
was created. This is currently used for parsing the view
when it is later queried. We use the same now when cloning
expressions from a view during condition pushdown.

Change-Id: I294e67e15b8fa45910afb60ade31b17127cfb572