Bug#36226080: wl16142-Mysqld failure-String::c_ptr_safe
Step 1: Cleanup of Item type classifications.

The reason for this failure is that Rapid prints item values according
to the type code. A stored procedure variable has no own item type code,
but rather impersonates as a literal value, with type code depending
on the data type of the variable. This means that a variable with string
type impersonates as STRING_ITEM, with integer type it impersonates as
INT_ITEM, etc.

The problem is then that a literal value is never NULL, whereas a
procedure variable can easily be NULL. Thus, printing a STRING_ITEM
does not check for NULL and when supplying a procedure variable, we'll
easily traverse a null pointer.

This is confusing at best and will be cleaned up. The first step is
to sanitize existing item type codes, remove those not in use and
rename a few for better understanding. The next step will add a specific
type code for variables declared inside routines (procedures, functions
and triggers).

- FIELD_AVG_ITEM, FIELD_STD_ITEM, FIELD_VARIANCE_ITEM and FIELD_BIT_ITEM
  are removed and replaced with one item code: AGGR_FIELD_ITEM.
- VARBIN_ITEM is renamed as HEX_BIN_ITEM, since it is used to
  represent hexadecimal and binary literal items.
- PROC_ITEM, XPATH_NODESET_CMP and VIEW_FIXER_ITEM are removed.
  XPATH_NODESET_CMP was deemed unnecessary and is now represented by
  FUNC_ITEM.
- METADATA_COPY_ITEM can be represented as TYPE_HOLDER_ITEM since it
  is only used locally when determining aggregated data types.
- SUBSELECT_ITEM is renamed as SUBQUERY_ITEM.
- TYPE_HOLDER is renamed as TYPE_HOLDER_ITEM.
- XPATH_NODESET is renamed as XPATH_NODESET_ITEM.

Change-Id: I934d5cb0283e43618b706715699cb46c44d9d141